test adding husky pre-commit hooks with lint
