+++ package.json
@@ -4,7 +4,7 @@
   "description": "REST express API to serve configs to magicbox-app",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "lint": "eslint src/**/*.js src/**/*.jsx"
   },
   "repository": {
     "type": "git",
@@ -27,7 +27,13 @@
     "chai-http": "^4.2.1",
     "eslint": "^5.16.0",
     "eslint-plugin-node": "^8.0.1",
+    "husky": "^1.3.1",
     "mocha": "^6.1.3",
     "nodemon": "^1.18.11"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "eslint src/**/*.js src/**/*.jsx"
+    }
   }
 }
