--- routes/index.js
@@ -1,11 +1,0 @@
-const express = require('express');
-// eslint-disable-next-line new-cap
-const router = express.Router();
-// const blobFetcher = require('../azure/blob-fetcher');
-// const helperShapefileCountry = require('../helpers/helper-shapefile-country')
-// const helperShapefile = require('../helpers/helper-shapefile')
-router.get('/', (req, res) => {
-  return res.send('hello world from index!');
-})
-
-module.exports = router;
